Add the ability to add HR data to pool swimming activities
This uses the interleaving technique of adding HR data as (interleaved)
RecordMesg entries. I was not successful at trying to implement the
HrMesg technique (of adding HR entries at the end of the fit file),
and since both Garmin Connect and Strava correctly interpet this
correctly, this will do for now.